Move renderer test libs to test folder
