+++ cobalt/shell/renderer/shell_content_renderer_client.cc
@@ -48,10 +48,9 @@
 #include "v8/include/v8.h"
 
 #if defined(RUN_BROWSER_TESTS)
-#include "cobalt/shell/common/shell_test_switches.h"            // nogncheck
-#include "cobalt/shell/renderer/shell_render_frame_observer.h"  // nogncheck
-#include "content/public/test/test_service.mojom.h"             // nogncheck
-#include "third_party/blink/public/web/web_testing_support.h"   // nogncheck
+#include "cobalt/shell/common/shell_test_switches.h"           // nogncheck
+#include "content/public/test/test_service.mojom.h"            // nogncheck
+#include "third_party/blink/public/web/web_testing_support.h"  // nogncheck
 #endif  // defined(RUN_BROWSER_TESTS)
 
 #if BUILDFLAG(ENABLE_PLUGINS)
@@ -254,13 +253,6 @@ void ShellContentRendererClient::ExposeInterfacesToBrowser(
 }
 
 #if defined(RUN_BROWSER_TESTS)
-void ShellContentRendererClient::RenderFrameCreated(RenderFrame* render_frame) {
-  // TODO(danakj): The ShellRenderFrameObserver is doing stuff only for
-  // browser tests. If we only create that for browser tests then the override
-  // of this method in WebTestContentRendererClient would not be needed.
-  new ShellRenderFrameObserver(render_frame);
-}
-
 void ShellContentRendererClient::DidInitializeWorkerContextOnWorkerThread(
     v8::Local<v8::Context> context) {
   if (base::CommandLine::ForCurrentProcess()->HasSwitch(
