+++ cobalt/shell/renderer/shell_content_renderer_client.h
@@ -44,7 +44,6 @@ class ShellContentRendererClient : public ContentRendererClient {
   void ExposeInterfacesToBrowser(mojo::BinderMap* binders) override;
 
 #if defined(RUN_BROWSER_TESTS)
-  void RenderFrameCreated(RenderFrame* render_frame) override;
   void DidInitializeWorkerContextOnWorkerThread(
       v8::Local<v8::Context> context) override;
 #endif  // defined(RUN_BROWSER_TESTS)
