+++ cobalt/testing/browser_tests/renderer/render_frame_test_helper.cc
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "cobalt/shell/renderer/render_frame_test_helper.h"
+#include "cobalt/testing/browser_tests/renderer/render_frame_test_helper.h"
 
 #include <utility>
