+++ cobalt/testing/browser_tests/renderer/shell_render_frame_observer.cc
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "cobalt/shell/renderer/shell_render_frame_observer.h"
+#include "cobalt/testing/browser_tests/renderer/shell_render_frame_observer.h"
 
 #include "base/command_line.h"
-#include "cobalt/shell/common/render_frame_test_helper.mojom.h"
 #include "cobalt/shell/common/shell_test_switches.h"
-#include "cobalt/shell/renderer/render_frame_test_helper.h"
+#include "cobalt/testing/browser_tests/common/render_frame_test_helper.mojom.h"
+#include "cobalt/testing/browser_tests/renderer/render_frame_test_helper.h"
 #include "content/public/renderer/render_frame.h"
 #include "content/public/renderer/render_frame_observer.h"
 #include "third_party/blink/public/web/web_testing_support.h"
